time_source

Trait TimeSource

pub trait TimeSource {
    // Required method
    fn get_current_time_millis(&self) -> u64;
}

A trait representing a source of time that can provide the current time in milliseconds.

Required Methods

fn get_current_time_millis(&self) -> u64

Returns the current time in milliseconds since Unix epoch.